% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypertension}
\alias{hypertension}
\title{Systolic blood pressure of 1653 patients.}
\format{
A data frame with 1653 rows and 20 variables:
\describe{
  \item{avgsysbp}{SBP, in mmHg}
  \item{D}{treatment assignment indicator, 1 if patient on Triamterene and ACEI, 0 if patient on ACEI alone}
  \item{PDC_avg}{1}
  \item{Studyid}{1}
  \item{approxAge2009}{1}
  \item{atrfib}{1}
  \item{avgdiasbp}{1}
  \item{black}{1}
  \item{bmicalc}{1}
  \item{cad}{1}
  \item{chf}{1}
  \item{ckd}{1}
  \item{copd}{1}
  \item{depression}{1}
  \item{diabetes}{1}
  \item{hyperlip}{1}
  \item{male}{1}
  \item{mi}{1}
  \item{pulse}{1}
  \item{stroke}{1}
}
}
\usage{
hypertension
}
\description{
A dataset containing the systolic blood pressure and other attributes of 1,653
patients.
}
\keyword{datasets}
