% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.rcate}
\alias{importance.rcate}
\title{Importance level from machine learning algorithms.}
\usage{
importance.rcate(object, plotit = TRUE, ...)
}
\arguments{
\item{object}{"rcate.ml" or "rcate.rf" object.}

\item{plotit}{whether plot the importance level.}

\item{...}{other.}
}
\value{
a list of components
\itemize{
 \item importance - vector of variable importance level.
 }
}
\description{
\code{importance.rcate} Returns the variable importance level.
}
