% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal.rcate.am}
\alias{marginal.rcate.am}
\title{Marginal treatment effect plot from additive model.}
\usage{
marginal.rcate.am(object, variable.col = 1, ...)
}
\arguments{
\item{object}{"rcate.am" object.}

\item{variable.col}{the column number of interested variable. Default is 1.}

\item{...}{other.}
}
\description{
\code{marginal.rcate.am} Returns the variable importance level from "rcate.am" model.
}
