% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal.rcate.ml}
\alias{marginal.rcate.ml}
\title{Marginal treatment effect plot from machine learning algorithms.}
\usage{
marginal.rcate.ml(object, variable.col = 1, ...)
}
\arguments{
\item{object}{"rcate.ml" object.}

\item{variable.col}{the column number of interested variable. Default is 1.}

\item{...}{other.}
}
\description{
\code{marginal.rcate.ml} Returns the variable importance level from "rcate.ml" model.
}
