% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.plot.R
\name{perm.plot.rcate.ml}
\alias{perm.plot.rcate.ml}
\title{Plot performance history of machine learning algorithm.}
\usage{
perm.plot.rcate.ml(object, ...)
}
\arguments{
\item{object}{"rcate.ml" object.}

\item{...}{other.}
}
\description{
\code{perm.plot.rcate.ml} Plot performance history of "rcate.ml" model.
}
