% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rcate.am}
\alias{predict.rcate.am}
\title{Predict treatment effect from additive model.}
\usage{
\method{predict}{rcate.am}(object, x, ...)
}
\arguments{
\item{object}{"RCATE" object.}

\item{x}{matrix or a data frame of predictors.}

\item{...}{other.}
}
\value{
a list of components
\itemize{
 \item predict - the robust estimation result of CATE.
 \item x - matrix of predictors.
 \item algorithm - fitting algorithm.
 \item model - "RCATE" object.
 \item method - estimation method.
 }
}
\description{
\code{predict} returns the robust treatment effect estimation result.
}
