% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rcate}
\alias{predict.rcate}
\title{Predict treatment effect.}
\usage{
\method{predict}{rcate}(object, x)
}
\arguments{
\item{object}{"RCATE" object.}

\item{x}{matrix or a data frame of predictors.}
}
\value{
a list of components
\itemize{
 \item predict - the robust estimation result of CATE.
 \item x - matrix of predictors.
 \item algorithm - fitting algorithm.
 \item model - "RCATE" object.
 \item method - estimation method.
 }
}
\description{
\code{predict} returns the robust treatment effect estimation result.
}
\details{
This is a generic function:
}
\examples{
n <- 1000; p <- 3
X <- matrix(runif(n*p,-3,3),nrow=n,ncol=p)
tau = sin(X[,1])
p = 1/(1+exp(-X[,1]))
d = rbinom(n,1,p)
t = 2*d-1
y = 1+tau*t/2 + rnorm(n,0,0.5)
fit <- rcate.ml(X,y,d)
y_pred <- predict.rcate(fit,X)$predict
}
