% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcate.rf.R
\name{predict.rcate.rf}
\alias{predict.rcate.rf}
\title{Predict treatment effect from robust random forests.}
\usage{
\method{predict}{rcate.rf}(object, newdata, ...)
}
\arguments{
\item{object}{a robust random forests.}

\item{newdata}{dataframe contains covariates.}

\item{...}{other.}
}
\value{
a list of components
\itemize{
 \item pred - prediction of newdata.
 \item newdata - a test data frame.
 }
}
\description{
\code{predict.rcate.rf} predicts treatment effect from robust random forests.
}
