% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcate.rf.R
\name{predict.reg.rf}
\alias{predict.reg.rf}
\title{Predict outcome from robust random forests.}
\usage{
\method{predict}{reg.rf}(object, newdata)
}
\arguments{
\item{object}{a robust random forests.}

\item{newdata}{dataframe contains covariates.}
}
\value{
a list of components
\itemize{
 \item pred - prediction of newdata.
 \item newdata - a test data frame.
 }
}
\description{
\code{predict.reg.rf} predicts outcome from robust random forests.
}
