% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcate.rf.R
\name{predict.reg.tree}
\alias{predict.reg.tree}
\title{Predict outcome from robust regression tree.}
\usage{
\method{predict}{reg.tree}(object, newdata)
}
\arguments{
\item{object}{a robust regression tree.}

\item{newdata}{dataframe contains covariates.}
}
\value{
a list of components
\itemize{
 \item pred - prediction of newdata.
 \item newdata - a test data frame.
 }
}
\description{
\code{predict.reg.tree} predicts outcome from robust regression tree.
}
