% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spline.R
\name{rcate.am}
\alias{rcate.am}
\title{Robust estimation of treatment effect using additive b-spline model.}
\usage{
rcate.am(
  x,
  y,
  d,
  method = "MCMEA",
  NIS = TRUE,
  nknots = NA,
  lambda.smooth = 2,
  nlambda = 30,
  nfolds = 5,
  n.trees.p = 40000,
  shrinkage.p = 0.005,
  n.minobsinnode.p = 10,
  interaction.depth.p = 1,
  cv.p = 2,
  n.trees.mu = c(1:50) * 50,
  shrinkage.mu = 0.01,
  n.minobsinnode.mu = 5,
  interaction.depth.mu = 5,
  cv.mu = 5
)
}
\arguments{
\item{x}{matrix or a data frame of predictors.}

\item{y}{vector of response values.}

\item{d}{vector of binary treatment assignment (0 or 1).}

\item{method}{character string of CATE estimation method: "MCMEA" - modified covariate
method with efficiency augmentation, "RL" - R-learning, or "DR" - doubly robust method.}

\item{NIS}{logical of using non-parametric independice screening or not. The default is TRUE
when $p < n/log(n)$.}

\item{nknots}{number of knots for cubic spline. The default is floor(sqrt(n)/2).}

\item{lambda.smooth}{scalar represent the smoothness penalty lambda2. The default is 2.}

\item{nlambda}{number of lambda1s for cross-validation. The default is 30.}

\item{nfolds}{number of folds for cross-validation. The default is 5.}

\item{n.trees.p}{tuning parameter the number of trees used for estimating propensity score
with GBM. the default value is 40000.}

\item{shrinkage.p}{tuning parameter the shrinkage level for estimating propensity score
with GBM. the default value is 0.005.}

\item{n.minobsinnode.p}{tuning parameter the minimum node size for estimating propensity score
with GBM. the default value is 10.}

\item{interaction.depth.p}{tuning parameter the number of interactions for estimating propensity
score with GBM. the default value is 1.}

\item{cv.p}{tuning parameter the number of folds in cross-validation for estimating propensity
score with GBM. the default value is 2.}

\item{n.trees.mu}{scalar or vector of the number of trees for estimating mean function with GBM.
The default is (1:50)*50.}

\item{shrinkage.mu}{tuning parameter the shrinkage level for estimating mean function
with GBM. the default value is 0.01.}

\item{n.minobsinnode.mu}{tuning parameter the minimum node size for estimating mean function
with GBM. the default value is 10.}

\item{interaction.depth.mu}{tuning parameter the number of interactions for estimating mean
function with GBM. the default value is 5.}

\item{cv.mu}{tuning parameter the folds for cross-validation for estimating mean function with
GBM. The default value is 5.}
}
\value{
a list of components
\itemize{
 \item model - the robust estimation model of CATE.
 \item method - estimation method.
 \item algorithm - fitting algorithm.
 \item lambda.smooth
 \item fitted.values - vector of fitted values.
 \item x - matrix of predictors.
 \item y - vector of response values.
 \item d - vector of treatment assignment.
 \item mean.x - column means of predictors.
 \item sd.x -column standard deviation of predictors.
 \item coef -coefficients.
 \item colnum - column number.
 \item nknots - number of knots of cubic spline.
 }
}
\description{
\code{rcate.am} returns the robust treatment effect estimation model.
}
\details{
This is a generic function:
}
\examples{
n <- 1000; p <- 3
X <- matrix(runif(n*p,-3,3),nrow=n,ncol=p)
tau = sin(X[,1])
p = 1/(1+exp(-X[,1]))
d = rbinom(n,1,p)
t = 2*d-1
y = 1+tau*t/2 + rnorm(n,0,0.5)
fit <- rcate.am(X,y,d)
y_pred <- predict.rcate(fit,X)$pred
}
