% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_rf.R
\name{rcate.rf}
\alias{rcate.rf}
\title{Robust estimation of treatment effect using random forests.}
\usage{
rcate.rf(
  x,
  y,
  d,
  method = "MCMEA",
  algorithm = "GBM",
  n.trees.p = 40000,
  shrinkage.p = 0.005,
  n.minobsinnode.p = 10,
  interaction.depth.p = 1,
  cv.p = 2,
  n.trees.mu = c(1:50) * 50,
  shrinkage.mu = 0.01,
  n.minobsinnode.mu = 5,
  interaction.depth.mu = 5,
  cv.mu = 5,
  n.trees.rf = 50,
  feature.frac = 1/2,
  newdata = NULL,
  minnodes = 5
)
}
\arguments{
\item{x}{matrix or a data frame of predictors.}

\item{y}{vector of response values.}

\item{d}{vector of binary treatment assignment (0 or 1).}

\item{method}{character string of CATE estimation method: "MCMEA" - modified covariate
method with efficiency augmentation, "RL" - R-learning, or "DR" - doubly robust method.}

\item{algorithm}{character string of algorithm: "GBM" - gradient boosting machine or
"NN" - neural network.}

\item{n.trees.p}{tuning parameter the number of trees used for estimating propensity score
with GBM. the default value is 40000.}

\item{shrinkage.p}{tuning parameter the shrinkage level for estimating propensity score
with GBM. the default value is 0.005.}

\item{n.minobsinnode.p}{tuning parameter the minimum node size for estimating propensity score
with GBM. the default value is 10.}

\item{interaction.depth.p}{tuning parameter the number of interactions for estimating propensity
score with GBM. the default value is 1.}

\item{cv.p}{tuning parameter the number of folds in cross-validation for estimating propensity
score with GBM. the default value is 2.}

\item{n.trees.mu}{scalar or vector of the number of trees for estimating mean function with GBM.
The default is (1:50)*50.}

\item{shrinkage.mu}{tuning parameter the shrinkage level for estimating mean function
with GBM. the default value is 0.01.}

\item{n.minobsinnode.mu}{tuning parameter the minimum node size for estimating mean function
with GBM. the default value is 10.}

\item{interaction.depth.mu}{tuning parameter the number of interactions for estimating mean
function with GBM. the default value is 5.}

\item{cv.mu}{tuning parameter the folds for cross-validation for estimating mean function with
GBM. The default value is 5.}

\item{n.trees.rf}{tuning parameter the number of trees used in GBM for estimating treatment
effect function if algorithm="GBM". The default is 1000.}

\item{feature.frac}{tuning parameter the number of interactions for estimating treatment
effect function if algorithm="GBM". The default value is 2.}

\item{newdata}{vector of the number of neurals in each hidden layer if algorithm='NN'.
The default is two layers with each layer the half size of previous layer.}

\item{minnodes}{vector of the dropout rate of each hidden layer if algorithm='NN'.
The default is no dropout.}
}
\value{
a list of components
\itemize{
 \item fit - estimation method.
 \item pred - prediction of newdata.
 }
}
\description{
\code{rcate.rf} returns the robust treatment effect estimation model.
}
\details{
This is a generic function:
}
\examples{
n <- 1000; p <- 3
X <- matrix(runif(n*p,-3,3),nrow=n,ncol=p)
tau = sin(X[,1])
p = 1/(1+exp(-X[,1]))
d = rbinom(n,1,p)
t = 2*d-1
y = 1+tau*t/2 + rnorm(n,0,0.5)
fit <- rcate.rf(X,y,d)
y_pred <- fit$pred
}
