% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcate.rf.R
\name{reg_rf}
\alias{reg_rf}
\title{Robust random forests.}
\usage{
reg_rf(formula, n_trees = 50, feature_frac = 1/2, data, weights, minnodes = 5)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{n_trees}{number of trees. Default is 50.}

\item{feature_frac}{fraction of features used in each split. Default is 1/2.}

\item{data}{a training data frame.}

\item{weights}{an optional vector of weights.}

\item{minnodes}{minimal leaf size of tree. Default is 5.}
}
\value{
a list of components
\itemize{
 \item fit - estimation method.
 \item importance - vector of variable importance.
 \item tree - tree structure and needed parameters.
 \item data - training data.
 \item nodepreds - leaf node predictions.
 }
}
\description{
\code{reg_rf} fit robust random forests based on weighted median splitting rule.
}
