% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcate.rf.R
\name{reg_tree_imp}
\alias{reg_tree_imp}
\title{Robust regression tree.}
\usage{
reg_tree_imp(formula, data, minsize, weights)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{data}{a training data frame.}

\item{minsize}{minimal leaf size of tree. Default is 5.}

\item{weights}{an optional vector of weights.}
}
\value{
a list of components
\itemize{
 \item tree - tree information.
 \item fit - estimation method.
 \item formula - an object of class "formula".
 \item importance - vector of importance level.
 \item data - a training data frame.
 \item pred - prediction of newdata.
 \item nodepred - leaf node prediction.
 }
}
\description{
\code{reg_tree} fit robust regression tree based on weighted median splitting rule.
}
