% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal.rcate}
\alias{marginal.rcate}
\title{Marginal treatment effect plot.}
\usage{
marginal.rcate(object, variable.name = NULL, ...)
}
\arguments{
\item{object}{"rcate.ml", "rcate.rf" or "rcate.am" object.}

\item{variable.name}{the name of interested variable. Default is the name of the
first continuous variable.}

\item{...}{other.}
}
\description{
\code{marginal.rcate} Returns the marginal treatment effect plot.
}
