% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rcate.am}
\alias{predict.rcate.am}
\title{Predict treatment effect from additive model.}
\usage{
\method{predict}{rcate.am}(object, x)
}
\arguments{
\item{object}{"RCATE" object.}

\item{x}{matrix or a data frame of predictors.}
}
\value{
a list of components
\itemize{
 \item predict - the robust estimation result of CATE.
 \item x - matrix of predictors.
 \item algorithm - fitting algorithm.
 \item model - "RCATE" object.
 \item method - estimation method.
 }
}
\description{
\code{predict} returns the robust treatment effect estimation result.
}
\examples{
n <- 1000; p <- 10
X <- matrix(rnorm(n*p,0,1),nrow=n,ncol=p)
tau = 6*sin(2*X[,1])+3*(X[,2])+X[,3]+9*tanh(0.5*X[,4])+3*X[,5]
p = 1/(1+exp(-X[,1]+X[,2]))
d = rbinom(n,1,p)
t = 2*d-1
y = 100+4*X[,1]+X[,2]-3*X[,3]+tau*t/2 + rnorm(n,0,1)
x_val = matrix(rnorm(200*10,0,1),nrow=200,ncol=10)
tau_val = 6*sin(2*x_val[,1])+3*(x_val[,2])+x_val[,3]+9*tanh(0.5*x_val[,4])+3*x_val[,5]

fit <- rcate.am(X,y,d)
y_pred <- predict.rcate.am(fit,x_val)$pred
plot(tau_val,y_pred);abline(0,1)
}
