% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rcate.am}
\alias{predict.rcate.am}
\title{Prediction of treatment effect from an \eqn{L_1} b-spline additive regression penalized model}
\usage{
\method{predict}{rcate.am}(object, x, ...)
}
\arguments{
\item{object}{"rcate.am" object.}

\item{x}{matrix or a data frame of predictors.}

\item{...}{other.}
}
\value{
a list of components
\itemize{
 \item predict - a robust estimation result of CATE.
 \item x - matrix of predictors.
 \item algorithm - fitting algorithm.
 \item model - "rcate.am" object.
 \item method - estimation method.
 }
}
\description{
\code{predict.rcate.am} Returns predicted treatment effect from "rcate.am" model.
}
