% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rcate.ml}
\alias{predict.rcate.ml}
\title{Predict treatment effect from machine learning algorithms.}
\usage{
\method{predict}{rcate.ml}(object, x, ...)
}
\arguments{
\item{object}{"rcate.ml" object.}

\item{x}{matrix or a data frame of predictors.}

\item{...}{other.}
}
\value{
a list of components
\itemize{
 \item predict - the robust estimation result of CATE.
 \item x - matrix of predictors.
 \item algorithm - fitting algorithm.
 \item model - "rcate.ml" object.
 \item method - estimation method.
 }
}
\description{
\code{predict.rcate.ml} Returns the predicted treatment effect from "rcate.ml" model.
}
