% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcate.rf.R
\name{reg_rf}
\alias{reg_rf}
\title{Robust random forests.}
\usage{
reg_rf(
  formula,
  n_trees = 50,
  feature_frac = 1/2,
  data,
  newdata,
  weights,
  minnodes = 5
)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{n_trees}{number of trees. Default is 50.}

\item{feature_frac}{fraction of features used in each split. Default is 1/2.}

\item{data}{a training data frame.}

\item{newdata}{an optional test data frame. If NULL, the newdata=data.}

\item{weights}{an optional vector of weights.}

\item{minnodes}{minimal leaf size of tree. Default is 5.}
}
\value{
a list of components
\itemize{
 \item fit - estimation method.
 \item pred - prediction of newdata.
 }
}
\description{
\code{reg_rf} fit robust random forests based on weighted median splitting rule.
}
\examples{
n <- 1000; p <- 3
X <- matrix(runif(n*p,-3,3),nrow=n,ncol=p)
y = 1+sin(X[,1]) + rnorm(n,0,0.5)
df <- data.frame(y,X)
RF <- reg_rf(y~X1+X2+X3,data=df,newdata=df,weights=rep(1,1000))
y_pred <- RF$pred
plot(y,y_pred)
}
