% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcate.rf.R
\name{reg_tree_imp}
\alias{reg_tree_imp}
\title{Robust regression tree.}
\usage{
reg_tree_imp(formula, data, minsize, newdata, weights)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{data}{a training data frame.}

\item{minsize}{minimal leaf size of tree. Default is 5.}

\item{newdata}{an optional test data frame. If NULL, the newdata=data.}

\item{weights}{an optional vector of weights.}
}
\value{
a list of components
\itemize{
 \item tree - tree information.
 \item fit - estimation method.
 \item formula - an object of class "formula".
 \item importance - vector of importance level.
 \item data - a training data frame.
 \item pred - prediction of newdata.
 \item newdata - a test data frame.
 }
}
\description{
\code{reg_tree} fit robust regression tree based on weighted median splitting rule.
}
\examples{
n <- 1000; p <- 3
X <- matrix(runif(n*p,-3,3),nrow=n,ncol=p)
y = 1+sin(X[,1]) + rnorm(n,0,0.5)
df <- data.frame(y,X)
tree <- reg_tree_imp(y~X1+X2+X3,data=df,minsize=3,newdata=df,weights=rep(1,1000))
y_pred <- tree$pred
plot(y,y_pred);abline(0,1)
}
